requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
torch
torchvision
pillow
google-cloud-storage
google-cloud-firestore
python-multipart
